runs fast now
but skinned mesh renderers don't work for some reason.
See line 128 on MeshRendererConnectorBase. it's throwing an error saying that Owner.AttachedGameObject is missing
Other than that, if this is fixed this is good for testing. still a bit jittery and trying to update head output between each frame causes a crash, so packets for now need to be proccessed as they come